make DuckDBNative class private
